SCRUM-214 refactor: Auth Domain 수정
