+++ src/main/java/Coordinate/coordikittyBE/domain/auth/login/controller/AuthController.java
@@ -1,20 +1,17 @@
 package Coordinate.coordikittyBE.domain.auth.login.controller;
 
-import Coordinate.coordikittyBE.domain.auth.login.dto.JwtTokenRequestDto;
-import Coordinate.coordikittyBE.domain.auth.login.dto.LoginResponseDto;
-import Coordinate.coordikittyBE.domain.auth.login.dto.LoginRequestDto;
-import Coordinate.coordikittyBE.domain.auth.login.dto.SocialLoginRequestDto;
+import Coordinate.coordikittyBE.domain.auth.login.dto.request.JwtTokenRequestDto;
+import Coordinate.coordikittyBE.domain.auth.login.dto.response.LoginResponseDto;
+import Coordinate.coordikittyBE.domain.auth.login.dto.response.LoginRequestDto;
+import Coordinate.coordikittyBE.domain.auth.login.dto.request.SocialLoginRequestDto;
 import Coordinate.coordikittyBE.domain.auth.login.service.TokenService;
 import Coordinate.coordikittyBE.domain.auth.login.service.UserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
 
 @RequiredArgsConstructor
 @RestController
