--- src/main/java/Coordinate/coordikittyBE/domain/auth/login/dto/JwtTokenRequestDto.java
@@ -1,5 +1,0 @@
-package Coordinate.coordikittyBE.domain.auth.login.dto;
-
-
-public record JwtTokenRequestDto(String refreshToken){
-}
