--- src/main/java/Coordinate/coordikittyBE/domain/auth/login/dto/SocialLoginRequestDto.java
@@ -1,4 +1,0 @@
-package Coordinate.coordikittyBE.domain.auth.login.dto;
-
-public record SocialLoginRequestDto(String email) {
-}
