+++ src/main/java/Coordinate/coordikittyBE/domain/auth/login/dto/request/JwtTokenRequestDto.java
@@ -1,0 +1,5 @@
+package Coordinate.coordikittyBE.domain.auth.login.dto.request;
+
+
+public record JwtTokenRequestDto(String refreshToken){
+}
