+++ src/main/java/Coordinate/coordikittyBE/domain/auth/login/dto/request/SocialLoginRequestDto.java
@@ -1,0 +1,4 @@
+package Coordinate.coordikittyBE.domain.auth.login.dto.request;
+
+public record SocialLoginRequestDto(String email) {
+}
