+++ src/main/java/Coordinate/coordikittyBE/domain/auth/login/dto/response/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package Coordinate.coordikittyBE.domain.auth.login.dto;
+package Coordinate.coordikittyBE.domain.auth.login.dto.response;
 
 
 public record LoginRequestDto(String email, String password) {
