+++ src/main/java/Coordinate/coordikittyBE/domain/auth/login/dto/response/LoginResponseDto.java
@@ -1,6 +1,7 @@
-package Coordinate.coordikittyBE.domain.auth.login.dto;
+package Coordinate.coordikittyBE.domain.auth.login.dto.response;
 
 import Coordinate.coordikittyBE.domain.auth.entity.User;
+import Coordinate.coordikittyBE.domain.auth.login.dto.TokenDto;
 
 public record LoginResponseDto(
         String email,
