+++ src/main/java/Coordinate/coordikittyBE/domain/auth/login/service/TokenService.java
@@ -2,7 +2,7 @@ package Coordinate.coordikittyBE.domain.auth.login.service;
 
 import Coordinate.coordikittyBE.domain.auth.entity.RefreshToken;
 import Coordinate.coordikittyBE.domain.auth.entity.User;
-import Coordinate.coordikittyBE.domain.auth.login.dto.LoginResponseDto;
+import Coordinate.coordikittyBE.domain.auth.login.dto.response.LoginResponseDto;
 import Coordinate.coordikittyBE.domain.auth.login.dto.TokenDto;
 import Coordinate.coordikittyBE.domain.auth.repository.UserRepository;
 import Coordinate.coordikittyBE.exception.CoordikittyException;
