+++ src/main/java/Coordinate/coordikittyBE/domain/auth/login/service/UserDetailServiceImpl.java
@@ -15,8 +15,8 @@ public class UserDetailServiceImpl implements UserDetailsService {
     private final UserRepository userRepository;
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        System.out.println(username);
-        return userRepository.findByName(username).orElseThrow(()-> new CoordikittyException(ErrorType.MEMBER_NOT_FOUND));
+        return userRepository.findByName(username)
+                .orElseThrow(()-> new CoordikittyException(ErrorType.MEMBER_NOT_FOUND));
     }
 
 }
