+++ src/main/java/Coordinate/coordikittyBE/exception/ErrorType.java
@@ -13,6 +13,7 @@ public enum ErrorType {
     TOKEN_PAYLOAD_EXTRACTION_FAILURE("AUTH-003", "토큰 페이로드 추출에 실패했습니다", 401),
     TOKEN_NOT_FOUND("AUTH-004", "토큰을 찾을 수 없습니다.", 401),
     REFRESH_TOKEN_ALREADY_EXIST("AUTH-005", "리프레시토큰이 이미 존재합니다.", 400),
+    INVALID_USER_ID("AUTH-006", "잘못된 유저 ID 입니다.", 401),
 
     MISSING_REQUIRED_VALUE_ERROR("COMMON-001", "필수 요청값이 누락되었습니다.", 400),
     INVALID_REQUEST_ERROR("COMMON-003", "올바르지 않은 데이터 요청입니다.", 400),
